After revision 1.4 of `src/usr.sbin/bsnmpd/bsnmpd/Makefile' the gensnmptree utility is required at "make depend" time, because it is used to generate `oid.h' for bsnmpd.
This means that all versions of HEAD before 2004/01/23 16:22:49 need
gensnmptree as a bootstrap util when it's not already installed as
part of the base system.  The first __FreeBSD_version that we can
assume this can work is 700014 which happened after the gensnmptree
change (in -r 1.263 of `src/sys/sys/param.h', at 2006/02/17 14:14:15).

For __FreeBSD_version values before 700014 add gensnmptree to the
bootstrap tools, to allow upgrades from versions of FreeBSD before
that date.

Approved by:	ru, harti


git-svn-id: svn+ssh://svn.freebsd.org/base/head@156493 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f