Removed some SMP ifdefs so that using the TSC as a cputime clock is not completely decided at config time.  Just don't default to using the TSC if there are multiple active CPUs.  Also, don't default to using the TSC if it is broken.  SMP ifdefs are still used to disallow using perfmon since perfmon is always broken if SMP is just configured.
This only helps much for SMP kernels running on 1 CPU.  The overheads
for using the i8254 cputime clock were a bit too high on 486/33's, and
now on multi-GHz CPUs they are usually in the 99-99.9% range.  Switching
from the old default of an i8254 clock to the TSC works poorly because
the overheads are not recalibrated.

Use the same condition for declaring perfmon stuff as for using it.