Remove direct ata_completed() call options from ata_finish(), except for the kernel dumping case.
ata_completed() may initiate ata_reinit() on error, that may lead to drives
attach or detach. Attach and detach are sending requests to drives and sleep
waiting for results. But ata_finish() can be called directly from
interrupt handler where sleeping is prohibited, so we must break this chain
somewhere. This place seems to fit best.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@189091 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f