MFC r213158:
Internalise reassembly queue related functionality and variables which should
not be used outside of the reassembly queue implementation. Provide a new
function to flush all segments from a reassembly queue and call it from the
appropriate places instead of manipulating the queue directly.

The base code from r213158 was modified as part of this MFC in order to work
correctly on FreeBSD 7.

Sponsored by:	FreeBSD Foundation
Reviewed by:	andre, gnn, rpaulo