Tab should not blank cells.
It seems I was under the impression that a tab differs from a single
forward tabulation, namely that it blanks the underlying cells. This
seems not to be the case. They are identical.

This should fix applications like jove(1) that use tabs instead of
explicit cursor position setting.

Reported by:	Brett Glass <brett lariat net>
MFC after:	3 days, after it's tested


git-svn-id: svn+ssh://svn.freebsd.org/base/head@226099 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f