According to the NetBSD foundation [1]:
Third parties are encouraged to change the license on any files which have
a 4-clause license contributed to the NetBSD Foundation to a 2-clause
license. We would also encourage you to inform us about these files, so
that we can continue to track the many places in which NetBSD is used.

http://www.netbsd.org/about/redistribution.html#why2clause [1]

Requested by:	joel@


git-svn-id: svn+ssh://svn.freebsd.org/base/head@226418 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f