- Change contigmalloc() to use the vm_paddr_t type instead of an unsigned   long for specifying a boundary constraint. - Change bus_dma tags to use bus_addr_t instead of bus_size_t for boundary   constraints.
These allow boundary constraints to be fully expressed for cases where
sizeof(bus_addr_t) != sizeof(bus_size_t).  Specifically, it allows a
driver to properly specify a 4GB boundary in a PAE kernel.

Note that this cannot be safely MFC'd without a lot of compat shims due
to KBI changes, so I do not intend to merge it.

Reviewed by:	scottl


git-svn-id: svn+ssh://svn.freebsd.org/base/head@232356 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f