Fix error messages containing the executed command name
Before, we took the first argument to pam_exec(8). With the addition of
options in front of the command, this could be wrong.

Now, options are parsed before calling _pam_exec() and messages contain
the proper command name.

While here, fix a warning.

Sponsored by:	Yakaz (http://www.yakaz.com)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@234184 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f