Some aac(4) adapters will always report that a direct access device is offline in response to a INQUIRY command that does not retreive vital product data(I personally have observed the behaviour on an Adaptec 2405 and a 5805).  Force the peripheral qualifier to "connected" so that upper layers correctly recognize that a disk is present.
This bug was uncovered by r216236.  Prior to that fix, aac(4) was
accidentally clearing the peripheral qualifier for all inquiry commands.

This fixes an issue where passthrough devices were not created for
disks behind aac(4) controllers suffering from the bug.  I have
verified that if a disk is not present that we still properly detect
that and not create the passthrough device.

Sponsored by:	Sandvine Incorporated
MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@240923 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f