Move snmp_hast manual to section 3, where all other manual pages for bsnmp modules are located.
Section 3 (Library Functions) looks wrong for this manual page, which
contains only module description, that is why initially it was located
to section 8 (System Manager's Manual).  On the other hand manual
pages for all other bsnmpd modules are already located in the section
3, and having all pages in one section looks more consistent.  Also,
currently, similarly to manuals for other modules, snmp_hast manual
contains LIBRARY section, which is not good style for section 8.

Requested by:	joel
X-MFC with:	r250379


git-svn-id: svn+ssh://svn.freebsd.org/base/head@250405 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f