Curb another regression (this one introduced by r251545, itself trying to fix the regression introduced by r251544; which was trying to make things consistent w/respect to ESC versus YES versus NO in the password disable prompt in "Login Management".
(need stronger coffee; very sorry for the churn)

With this revision, the "YES", "NO", and implied ESC options all work as-
expected. Choosing "YES" allows you to proceed and the password will be
disabled. Choosing "NO" will bring back around to enter a password for the
account. Pressing ESC will drop you out of either user or group input and
back to the usermgmt screen.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@251547 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f