dd a 24 bits wide ext_flags field to m_ext by reducing ext_type to 8 bits.  ext_type is an enumerator and the number of types we have is a mere dozen.
A couple of ext_types are renumbered to fit within 8 bits.

EXT_VENDOR[1-4] and EXT_EXP[1-4] types for vendor-internal and
experimental local mapping.

The ext_flags field is currently unused but has a couple of flags
already defined for future use.  Again vendor and experimental
flags are provided for local mapping.

EXT_FLAG_BITS is provided for the printf(9) %b identifier.

Initialize and copy ext_flags in the relevant mbuf functions.

Improve alignment and packing of struct m_ext on 32 and 64 archs
by carefully sorting the fields.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@254780 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f