Change the behavior of bhyve such that the gdb listening port is opt-in rather than opt-out.
Prior to this change if the "-g" option was not specified then a listening
socket for tunneling gdb packets would be opened at port 6466. If a second
virtual machine is fired up, also without the "-g" option, then that would
fail because there is already a listener on port 6466.

After this change if a gdb tunnel port needs to be created it needs to be
explicitly specified with a "-g <portnum>" command line option.

Reviewed by:	grehan@
Approved by:	re@ (blanket)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@256156 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f