ssh-copy-id: avoid sending private keys; add -v option
To help avoid confusion: when attempting to send a key file check to see if a
file of the same name exists with a '.pub' suffix and send that instead.  This
mimics the behavior of other ssh-copy-id scripts.

Add -v passthrough.

Reported by:	dweimer <dweimer@dweimer.net>
Reported by:	feld
MFC After:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@262645 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f