loader: fix the size of MODINFOMD_MODULEP
The data in MODINFOMD_MODULEP is packed by the loader as a 4 byte type, but
the amd64 kernel expects a vm_paddr_t, which is of size 8 bytes. Fix this by
saving it as 8 bytes in the loader and retrieving it using the proper type
in the kernel.

Sponsored by: Citrix Systems R&D


git-svn-id: svn+ssh://svn.freebsd.org/base/head@277418 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f