Overhaul the relnotes page to reorganize the sections.
With the previous layout, the majority of the release
notes were categorized under one main "What's New"
section, which in addition to making the page difficult
to parse relevant bits, also had a few rendering issues,
such as 'Note' blocks visually not appearing as if they
pertain to a particular change.

This change removes the "What's New" section, integrating
it with the introduction.  In addition, each section now
provides a brief description of its contents, broken down
further into a subsection for each "topic", making it much
easier to find a specific topic in the page.

MFC after:	3 days
X-MFC-Note:	direct commit, not a merge
Sponsored by:	The FreeBSD Foundation