Add a new flag, SBUF_INCLUDENUL, and new get/set/clear functions for flags.
The SBUF_INCLUDENUL flag causes the nulterm byte at the end of the string
to be counted in the length of the data.  If copying the data using the
sbuf_data() and sbuf_len() functions, or if writing it automatically with
a drain function, the net effect is that the nulterm byte is copied along
with the rest of the data.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@279992 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f