Raise the SDHCI timeout to 10 seconds and add a sysctl to allow changing this value at runtime.
The SD card specification says that a block write or a block erase can take
up to 250ms to complete and thus, under some circumstances, the existent 2
seconds timeout was triggering with normal usage.

This change fixes the sporadic controller timeout that happens on RPi and
RPi 2.

Discussed with:		ian (some time ago)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@283263 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f